separate navbar and SearchBar components
